First upload.
This is the initial version of the single page site for Miss Chiff. It
has a simple navigation bar at the bottom of the page. Additionally,
there is a scoll function where bio info animates on screen after
scroll hits the bottom of the page.

Marilyn want to omit this functionality and bio altogether.

She also wants to change the font to a font-squirel webfont.